requirements.txt
python-telegram-bot==21.3
asyncpg==0.29.0
httpx==0.27.0
python-dotenv==1.0.1
